extract env.handling from main-menu view
