+++ packages/server/src/server.ts
@@ -1,33 +1,30 @@
+import { prometheus } from "@hono/prometheus";
 import { Server } from "bun";
-import { makeFrontpageUrl } from "decorator-shared/urls";
+import { CsrPayload } from "decorator-shared/types";
 import { Hono } from "hono";
 import { serveStatic } from "hono/bun";
 import { HTTPException } from "hono/http-exception";
 import { cspDirectives } from "./content-security-policy";
+import { csrAssets } from "./csr";
 import { env } from "./env/server";
 import { authHandler } from "./handlers/auth-handler";
 import { headers } from "./handlers/headers";
 import { searchHandler } from "./handlers/search-handler";
 import { versionProxyHandler } from "./handlers/version-proxy";
-import i18n from "./i18n";
-import { getMainMenuLinks, mainMenuContextLinks } from "./menu/main-menu";
+import { headAssets, HeadAssetsTemplate } from "./head";
 import { setupMocks } from "./mocks";
 import { archiveNotification } from "./notifications";
 import { fetchOpsMessages } from "./ops-msgs";
 import { getTaskAnalyticsConfig } from "./task-analytics-config";
 import { getFeatures } from "./unleash";
+import { isLocalhost } from "./urls";
 import { validParams } from "./validateParams";
 import { IndexTemplate } from "./views";
-import { MainMenu, MainMenuTemplate } from "./views/header/main-menu";
-import { prometheus } from "@hono/prometheus";
-import { HeaderTemplate } from "./views/header/header";
 import { FooterTemplate } from "./views/footer/footer";
+import { HeaderTemplate } from "./views/header/header";
+import { MainMenuTemplate } from "./views/header/render-main-menu";
 import { buildDecoratorData, ScriptsTemplate } from "./views/scripts";
 import { StylesTemplate } from "./views/styles";
-import { csrAssets } from "./csr";
-import { CsrPayload } from "decorator-shared/types";
-import { headAssets, HeadAssetsTemplate } from "./head";
-import { isLocalhost } from "./urls";
 
 const startupTime = Date.now();
